src/main/webapp/app/entities/pigeon/delete/pigeon-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { PigeonService } from '../service/pigeon.service';

import { PigeonDeleteDialogComponent } from './pigeon-delete-dialog.component';

describe('Pigeon Management Delete Component', () => {
  let comp: PigeonDeleteDialogComponent;
  let fixture: ComponentFixture<PigeonDeleteDialogComponent>;
  let service: PigeonService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, PigeonDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(PigeonDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(PigeonDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(PigeonService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
